PP-1224 Refactoring Refund Model package name
with @rahul-vinayak
