+++ src/main/java/uk/gov/pay/api/model/RefundFromConnector.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.api.resources;
+package uk.gov.pay.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
